% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cruza Dihi.R
\name{Cruza_Dihi}
\alias{Cruza_Dihi}
\title{Cruzamiento Dihibrido}
\usage{
Cruza_Dihi(Macho, Hembra, c_dominante1, c_dominante2, c_recesiva1, c_recesiva2)
}
\arguments{
\item{Macho}{(vector) variable que representa los alelos del macho en la cruza.}

\item{Hembra}{(vector) variable que representa los alelos de la hembra en la cruza.}

\item{c_dominante1}{(vector) Variable que representa elprimer alelo dominante (primera caracteristica).}

\item{c_dominante2}{(vector) Variable que representa el segundo alelo dominante (segunda caracteristica).}

\item{c_recesiva1}{(vector) Variable que representa el primer alelo recesivo (primera caracteristica).}

\item{data}{base de datos}

\item{c_drecesiva2}{(vector) Variable que representa el segundo alelo recesivo (segunda caracteristica).}
}
\value{
una lista con una matriz de las distintas combinaciones de alelos, una tabla de proorciones genotipicas, una matriz de todas las coombinaciones posibles en cuanto a las caracteristicas fenotipicas y una tabla de proporci贸n fenotipica.
}
\description{
El prop贸sito de esta funci贸n es determinar si existe alguna relaci贸n entre los pares alelicos para deteminar 2 caracteristicas (genotipicas y fenotipicas) en sus descendientes.
}
