% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CruzaMon.R
\name{Cruza_Mon}
\alias{Cruza_Mon}
\title{Cruzamiento Monohibrido
Determina en un método simple las proporciones teóricas de la expresión genética que se presenta en la progenie, de a cuerdo a las caracteristicas de los padres.}
\usage{
Cruza_Mon(Macho, Hembra, c_dominate, c_recesiva)
}
\arguments{
\item{Macho}{(vector)variable que representa los alelos del macho en la cruza.}

\item{Hembra}{(vector) variable que representa los alelos de la hembra en la cruza.}

\item{c_recesiva}{(vector) variable que representa una caracteristica recesiva en la cruza.}

\item{data}{base de datos}

\item{c_dominante}{(vector) variable que representa una caracteristica dominante en la cruza.}
}
\value{
una lista con una matriz con la cruza y posibles combinaciones genotipicas, proporciones genotipicas, una matriz con las posibles caracteristcas fenotipicas y proporciones fenotipicas.
}
\description{
Cruzamiento Monohibrido
Determina en un método simple las proporciones teóricas de la expresión genética que se presenta en la progenie, de a cuerdo a las caracteristicas de los padres.
}
