% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TablaChicuaG.R
\name{Tablachi}
\alias{Tablachi}
\title{Tabla de chi cuadrada}
\usage{
Tablachi(Fenotipos, tot_Crias, valor_o, proporción)
}
\arguments{
\item{Fenotipos}{(vector) variable que presenta las caracteristicas fenotipcas de acuerdo a la proporcion a estudiar.}

\item{tot_Crias}{(vector) variable que indica el total de crías a estudiar en una generación.}

\item{valor_o}{(vector) Variable que expresa los datos fenotipicos observados en el total de crías.}

\item{proporción}{(Vector) variable que representa la proporcion de cada fenotipo.}

\item{data}{base de datos}
}
\value{
una lista que contiene una tabla de analisis de chi cuadrada, la hipotesis que se aprueba.
}
\description{
Comprueba hipótesis de ciertos datos geneticos y acepta o descarta si son como se esperaba, compara valores observados con los datos esperados que se tendrian si la hipotesis nula es cierta.
}
